ingest/pyproject.toml
[tool.poetry]
name = "ingest"
version = "0.1.0"
description = "Ingest service for SproutSense plant monitoring system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
flask = "^3.1.0"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ingest = "ingest.main:main"
